Convert H5PEditor to Typescript
